+++ barelymusician/common/random.cpp
@@ -11,15 +11,15 @@ namespace barely {
 
 Random::Random(int seed) noexcept : engine_(static_cast<unsigned int>(seed)) {}
 
-double Random::DrawNormal(double mean, double variance) noexcept {
+float Random::DrawNormal(float mean, float variance) noexcept {
   assert(variance >= 0.0);
-  const std::normal_distribution<double>::param_type param(mean, variance);
+  const std::normal_distribution<float>::param_type param(mean, variance);
   return normal_distribution_(engine_, param);
 }
 
-double Random::DrawUniform(double min, double max) noexcept {
+float Random::DrawUniform(float min, float max) noexcept {
   assert(min <= max);
-  const std::uniform_real_distribution<double>::param_type param(min, max);
+  const std::uniform_real_distribution<float>::param_type param(min, max);
   return uniform_real_distribution_(engine_, param);
 }
