+++ barelymusician/dsp/decibels.cpp
@@ -5,18 +5,18 @@
 
 namespace barely {
 
-double AmplitudeFromDecibels(double decibels) noexcept {
+float AmplitudeFromDecibels(float decibels) noexcept {
   if (decibels > kMinDecibels) {
     // amplitude = 10 ^ (decibels / 20).
-    return std::pow(10.0, 0.05 * decibels);
+    return std::pow(10.0f, 0.05f * decibels);
   }
-  return 0.0;
+  return 0.0f;
 }
 
-double DecibelsFromAmplitude(double amplitude) noexcept {
-  if (amplitude > 0.0) {
+float DecibelsFromAmplitude(float amplitude) noexcept {
+  if (amplitude > 0.0f) {
     // decibels = 20 * log(amplitude).
-    return 20.0 * std::log10(amplitude);
+    return 20.0f * std::log10(amplitude);
   }
   return kMinDecibels;
 }
