+++ barelymusician/dsp/gain_processor.h
@@ -16,25 +16,25 @@ class GainProcessor {
   /// @param buffer Buffer.
   /// @param channel_count Number of channels.
   /// @param frame_count Number of frames.
-  void Process(double* buffer, int channel_count, int frame_count) noexcept;
+  void Process(float* buffer, int channel_count, int frame_count) noexcept;
 
   /// Sets gain.
   ///
   /// @param gain Gain in amplitude.
-  void SetGain(double gain) noexcept;
+  void SetGain(float gain) noexcept;
 
  private:
   // Current gain in amplitude.
-  double gain_ = 1.0;
+  float gain_ = 1.0f;
 
   // Denotes whether processor started processing or not.
   bool is_initialized_ = false;
 
   // Total number of ramp frames for unity gain.
-  double unity_ramp_frame_count_ = 0.0;
+  float unity_ramp_frame_count_ = 0.0f;
 
   // Target gain in amplitude.
-  double target_gain_ = 1.0;
+  float target_gain_ = 1.0f;
 };
 
 }  // namespace barely
