+++ barelymusician/dsp/sample_player.h
@@ -14,7 +14,7 @@ class SamplePlayer : public Generator {
   explicit SamplePlayer(int frame_rate) noexcept;
 
   /// Implements `Generator`.
-  double Next() noexcept override;
+  float Next() noexcept override;
   void Reset() noexcept override;
 
   /// Sets the sample data.
@@ -22,7 +22,7 @@ class SamplePlayer : public Generator {
   /// @param data Sample data.
   /// @param frequency Data sampling frequency in hertz.
   /// @param length Data length in samples.
-  void SetData(const double* data, int frequency, int length) noexcept;
+  void SetData(const float* data, int frequency, int length) noexcept;
 
   /// Sets whether the playback should be looping.
   ///
@@ -32,32 +32,32 @@ class SamplePlayer : public Generator {
   /// Sets the playback speed.
   ///
   /// @param speed Playback speed.
-  void SetSpeed(double speed) noexcept;
+  void SetSpeed(float speed) noexcept;
 
  private:
   // Calculates per sample increment amount with the current properties.
   void CalculateIncrementPerSample() noexcept;
 
   // Inverse frame rate in seconds.
-  double frame_interval_ = 0.0;
+  float frame_interval_ = 0.0f;
 
   // Sample data.
-  const double* data_ = nullptr;
+  const float* data_ = nullptr;
 
   // Sample data frame rate in hertz.
-  double frequency_ = 0.0;
+  float frequency_ = 0.0f;
 
   // Sample data length in samples.
-  double length_ = 0.0;
+  float length_ = 0.0f;
 
   // Playback speed.
-  double speed_ = 1.0;
+  float speed_ = 1.0f;
 
   // Playback cursor.
-  double cursor_ = 0.0;
+  float cursor_ = 0.0f;
 
   // Increment per sample.
-  double increment_ = 0.0;
+  float increment_ = 0.0f;
 
   // Denotes whether the playback is looping or not.
   bool loop_ = false;
